chore: drop unused files
